feat(querysvc): Address second round of review feedback
- Move FindTraceSummaries method into service.go alongside other QueryService methods
- Move findTraceSummaries HTTP handler back to http_gateway.go; summaries.go now
  holds only the pre-IDL JSON types
- Apply clock-skew adjuster before summarizing in the fallback path, consistent
  with FindTraces behavior
- Guard zero StartTime in JSON response (emit 0 instead of large negative value)
- Rename local var stats → svcStats per reviewer suggestion
